made style changes to some components
