+++ frontend/my-bookshelf/src/Goals.css
@@ -1,10 +1,19 @@
 .goals-button {
     margin-top: 250px;
-    margin-left: 150px;
     display: flex;
     justify-content: center;
+    padding: 10px 20px;
+    font-size: 16px;
+    cursor: pointer;
+}
 
-
+.goal-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin-top: 60px;
+    gap: 20px;
 }
 
 #goals td,
@@ -22,4 +31,15 @@
 .is-public {
     width: 20px;
     height: 20px;
+}
+
+.back-arrow {
+    position: absolute;
+    top: 20px;
+    left: 20px;
+}
+
+#goBack {
+    font-size: 50px;
+    cursor: pointer;
 }
