+++ frontend/my-bookshelf/src/ReflectionPage.css
@@ -20,3 +20,30 @@
     background-color: #CBD5C0;
     color: #fff;
 }
+
+.reflection-page {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+    margin: 0;
+    border: 0;
+
+}
+
+.reflection-header {
+    width: 500px;
+    border: 0;
+    margin: 0 30px;
+    flex-shrink: 0;
+}
+
+.review-reflection {
+    margin: 0;
+    padding: 0;
+}
+
+.wrapper {
+    max-width: 1200px;
+    margin: 0 auto;
+    padding: 0 20px;
+}
