+++ frontend/my-bookshelf/src/ReflectionPage.jsx
@@ -98,32 +98,34 @@ function ReflectionPage(props) {
         return () => { window.removeEventListener('REFLECTION_SAVED', handleSync) }
     }, [bookData.googleId])
     return (
-        <>
-            <div>
-                <img src={bookData.cover} alt="bookcover" />
-                <h3>{bookData.title} - {bookData.authors.join(", ")}</h3>
-                <p>{bookData.description}</p>
-                <div className="buy-links">
-                    <a href={`https://www.barnesandnoble.com/s/${encodeURIComponent(bookData.title + ' ' + bookData.authors.join(", "))}`}>Buy on Barnes & Noble</a>
-                    <a href={` https://www.amazon.com/s?k=${encodeURIComponent(bookData.title + ' ' + bookData.authors.join(", "))}`}>Buy on Amazon</a>
+        <div className="wrapper">
+            <div className="reflection-page">
+                <div className="reflection-header">
+                    <img width="200px" height="300px" src={bookData.cover} alt="bookcover" />
+                    <h3>{bookData.title} - {bookData.authors.join(", ")}</h3>
+                    <p>{bookData.description}</p>
+                    <div className="buy-links">
+                        <a href={`https://www.barnesandnoble.com/s/${encodeURIComponent(bookData.title + ' ' + bookData.authors.join(", "))}`}>Buy on Barnes & Noble</a>
+                        <a href={` https://www.amazon.com/s?k=${encodeURIComponent(bookData.title + ' ' + bookData.authors.join(", "))}`}>Buy on Amazon</a>
+                    </div>
                 </div>
-            </div>
-            <ReviewForm setContent={setContent} setRating={setRating} content={content} rating={rating} reviews={reviews} getReviews={getReviews} bookData={bookData} />
-            <ReviewsPage etContent={setContent} setRating={setRating} content={content} rating={rating} reviews={reviews} getReviews={getReviews} bookData={bookData} />
-            {existingReflection && !editMode ? (
-                <>
-                    <p><strong>Your Reflection:</strong></p>
-                    <p>{reflection}</p>
-                    <button onClick={() => setEditMode(true)}><i className="fa-solid fa-pen"></i></button></>
+                <div className="review-reflection">
+                    <ReviewForm setContent={setContent} setRating={setRating} content={content} rating={rating} reviews={reviews} getReviews={getReviews} bookData={bookData} />
+                    <ReviewsPage etContent={setContent} setRating={setRating} content={content} rating={rating} reviews={reviews} getReviews={getReviews} bookData={bookData} />
+                    {existingReflection && !editMode ? (
+                        <>
+                            <p><strong>Your Reflection:</strong></p>
+                            <p>{reflection}</p>
+                            <button onClick={() => setEditMode(true)}><i className="fa-solid fa-pen"></i></button></>
 
-            ) : (<div>
+                    ) : (<div>
 
-                <form action="">
-                    <textarea value={reflection} onChange={(e) => setReflection(e.target.value)} rows="40" cols="40" name="" id=""></textarea>
-                </form>
-                <button onClick={handleSave}>{existingReflection ? "Save Changes" : "Save Reflection"}</button>
-            </div>)
-            }</>
+                        <form action="">
+                            <textarea value={reflection} onChange={(e) => setReflection(e.target.value)} rows="40" cols="40" name="" id=""></textarea>
+                        </form>
+                        <button onClick={handleSave}>{existingReflection ? "Save Changes" : "Save Reflection"}</button>
+                    </div>)
+                    }</div></div></div>
 
     )
 }
