+++ frontend/my-bookshelf/src/ReviewForm.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { useUser } from "./contexts/UserContext";
 import { addToQueue } from "./utils/db";
 import { toast } from "react-toastify";
-
+import StarRatingInput from "./StarRatingInput";
 function ReviewForm(props) {
     const [review, setReview] = useState("");
     const [rating, setRating] = useState(0);
@@ -66,34 +66,6 @@ function ReviewForm(props) {
             window.removeEventListener("REVIEW_SAVED", handleSync);
         };
     }, [props.bookData.googleId]);
-
-    // Star rating component inside ReviewForm for simplicity
-    const StarRatingInput = ({ rating, setRating }) => {
-        return (
-            <div>
-                {[1, 2, 3, 4, 5].map((star) => (
-                    <i
-                        key={star}
-                        className={`fa-star fa${star <= rating ? "s" : "r"}`}
-                        style={{
-                            cursor: "pointer",
-                            color: star <= rating ? "gold" : "gray",
-                            marginRight: 5,
-                            fontSize: 24,
-                        }}
-                        onClick={() => setRating(star)}
-                        aria-label={`${star} star`}
-                        role="button"
-                        tabIndex={0}
-                        onKeyDown={(e) => {
-                            if (e.key === "Enter" || e.key === " ") setRating(star);
-                        }}
-                    ></i>
-                ))}
-            </div>
-        );
-    };
-
     return (
         <div>
             {!userReview ? (
@@ -112,7 +84,7 @@ function ReviewForm(props) {
                     <button type="submit">Submit</button>
                 </form>
             ) : (
-                <div>Review submitted</div>
+                <div></div>
             )}
         </div>
     );
